renamed impl to data files and layout fixes added
