+++ src/components/button/button-template.ts
@@ -1,20 +1,22 @@
-const buttonTemplate: string = `<button
-  {{#if className}}
-    class="{{className}}"
-  {{/if}}
-
-  {{#if id}}
-    id='{{id}}'
-  {{/if}}
-
-  {{#if type}}
-    id='{{type}}'
-  {{/if}}
-
-  {{#if form}}
-    form='{{form}}'
-  {{/if}}
-
->{{buttonText}}</button>`
+const buttonTemplate: string = `
+    <button
+      {{#if className}}
+        class="{{className}}"
+      {{/if}}
+    
+      {{#if id}}
+        id='{{id}}'
+      {{/if}}
+    
+      {{#if type}}
+        type='{{type}}'
+      {{/if}}
+    
+      {{#if form}}
+        form='{{form}}'
+      {{/if}}
+    
+    >{{buttonText}}</button>
+`
 
 export default buttonTemplate
