+++ src/pages/404/index.ts
@@ -3,7 +3,7 @@ import Block from '../../modules/block'
 import './not-found.scss'
 import prepareTemplate from '../../modules/prepare-template'
 import { PageNotFoundProps as PageProps } from './not-found-model'
-import pageData from './not-found-impl'
+import pageData from './not-found-data'
 
 export default class PageNotFound extends Block {
   constructor(props: PageProps) {
