+++ src/pages/404/not-found-data.ts
