+++ src/pages/404/not-found-template.ts
@@ -1,9 +1,11 @@
-const notFoundTemplate = `<div class='align-center'>
-  <div class='wrapper'>
-    <h2>{{description}}</h2>
-    <p>{{title}}</p>
-    <label>{{{button}}}</label>
-  </div>
-</div>`
+const notFoundTemplate = `
+    <div class='align-center'>
+      <div class='wrapper'>
+        <h2>{{description}}</h2>
+        <p>{{title}}</p>
+        {{{button}}}
+      </div>
+    </div>
+`
 
 export default notFoundTemplate
