+++ src/pages/500/server-error-template.ts
@@ -1,9 +1,11 @@
-const serverErrorTemplate = `<div class='align-center'>
-  <div class='wrapper'>
-    <h2>{{description}}</h2>
-    <p>{{title}}</p>
-    <label>{{{button}}}</label>
-  </div>
-</div>`
+const serverErrorTemplate = `
+    <div class='align-center'>
+      <div class='wrapper'>
+        <h2>{{description}}</h2>
+        <p>{{title}}</p>
+        {{{button}}}
+      </div>
+    </div>
+`
 
 export default serverErrorTemplate
