+++ src/pages/register/register-data.ts
