+++ src/pages/register/register.scss
@@ -12,7 +12,7 @@
     .buttons-block {
       position: relative;
       bottom: 0;
-      padding: 0 30px;
+      padding: 0;
       left: 0;
       right: 0;
       top: 20px;
