+++ src/styles/_layout.scss
@@ -27,8 +27,6 @@ body {
 
     input {
       width: 100%;
-      /* border-color: $formInputBorderColor;
-      border-width: 0 0 1px 0; */
       border: none;
       background: initial;
       outline: none;
